refactor(x/swingset): Move queue abstraction to vstorage
